+++ pybamm/models/submodels/interface/base_interface.py
@@ -339,8 +339,8 @@ class BaseInterface(pybamm.BaseSubModel):
             }
         )
 
-        a_n = variables["Negative electrode surface area per unit volume"]
-        a_p = variables["Positive electrode surface area per unit volume"]
+        a_n = variables["Negative electrode surface area to volume ratio"]
+        a_p = variables["Positive electrode surface area to volume ratio"]
         a = pybamm.Concatenation(
             a_n, pybamm.FullBroadcast(0, "separator", "current collector"), a_p
         )