+++ tests/unit/test_models/test_submodels/test_electrolyte_conductivity/test_full_conductivity.py
@@ -11,7 +11,7 @@ class TestFull(unittest.TestCase):
     def test_public_functions(self):
         param = pybamm.LithiumIonParameters()
         a = pybamm.Scalar(0)
-        surf = "electrode surface area per unit volume"
+        surf = "electrode surface area to volume ratio"
         variables = {
             "Electrolyte tortuosity": a,
             "Electrolyte concentration": pybamm.FullBroadcast(